[validate] - Improve unit testing

+++ validate/validate_test.go
@@ -2,15 +2,19 @@
 package validate
 
 import (
+	"context"
 	"os"
 	"os/exec"
 	"os/signal"
 	"testing"
 	"time"
 
+	"github.com/Lord-Y/synker/commons"
 	"github.com/Lord-Y/synker/elasticsearch"
 	"github.com/Lord-Y/synker/kafka"
 	"github.com/Lord-Y/synker/tools"
+	"github.com/icrowley/fake"
+	"github.com/jackc/pgx/v4/pgxpool"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -144,19 +148,24 @@ func TestManageTopicsAndElasticsearchIndex(t *testing.T) {
 		client, err := elasticsearch.Client()
 		assert.Nil(err)
 
-		b, err := elasticsearch.DeleteIndex(client, "user_promo_codes")
+		indexExist, err := elasticsearch.IndexAlreadyExist(client, "user_promo_codes")
 		assert.Nil(err)
-		assert.Equal(true, b)
+		if indexExist {
+			client, err := elasticsearch.Client()
+			assert.Nil(err)
+			b, err := elasticsearch.DeleteIndex(client, "user_promo_codes")
+			assert.Nil(err)
+			assert.Equal(true, b)
+		}
+		var c Validate
+		c.ConfigDir = "examples/schemas"
+		c.Run()
+
+		err = c.ManageTopics()
+		assert.Nil(err)
+		err = c.ManageElasticsearchIndex()
+		assert.Nil(err)
 	}
-
-	var c Validate
-	c.ConfigDir = "examples/schemas"
-	c.Run()
-
-	err = c.ManageTopics()
-	assert.Nil(err)
-	err = c.ManageElasticsearchIndex()
-	assert.Nil(err)
 }
 
 func TestManageTopicsAndElasticsearchIndex_with_alias(t *testing.T) {
@@ -308,6 +317,7 @@ func TestValidate_processing(t *testing.T) {
 		c.ConfigDir = "examples/schemas"
 		c.Run()
 		c.Processing()
+		time.Sleep(10 * time.Second)
 		signal.Stop(sigc)
 	}()
 
@@ -316,4 +326,66 @@ func TestValidate_processing(t *testing.T) {
 		t.Fatal(err)
 	}
 	time.Sleep(1 * time.Second)
+}
+
+func TestValidate_processing_with_new_id(t *testing.T) {
+	assert := assert.New(t)
+
+	proc, err := os.FindProcess(os.Getpid())
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	sigc := make(chan os.Signal, 1)
+	signal.Notify(sigc, os.Interrupt)
+
+	go func() {
+		<-sigc
+		var c Validate
+		c.ConfigDir = "examples/schemas"
+		c.Run()
+		err = insertIntoUserPromoCodesForUnitTesting()
+		assert.Nil(err)
+		c.Processing()
+		signal.Stop(sigc)
+	}()
+
+	err = proc.Signal(os.Interrupt)
+	if err != nil {
+		t.Fatal(err)
+	}
+	time.Sleep(1 * time.Second)
+}
+
+func insertIntoUserPromoCodesForUnitTesting() (err error) {
+	ctx := context.Background()
+	db, err := pgxpool.Connect(ctx, commons.GetPGURI())
+	if err != nil {
+		return
+	}
+	defer db.Close()
+
+	tx, err := db.Begin(ctx)
+	if err != nil {
+		return
+	}
+	//golangci-lint fail on this check while the transaction error is checked
+	defer tx.Rollback(ctx) //nolint
+
+	_, err = tx.Exec(
+		ctx,
+		"INSERT INTO user_promo_codes VALUES($1,$2,$3,NOW(),$4)",
+		"amsterdam",
+		"ae147ae1-47ae-4800-8000-000000000022",
+		fake.CharactersN(5),
+		10,
+	)
+	if err != nil {
+		return
+	}
+
+	if err = tx.Commit(ctx); err != nil {
+		return
+	}
+	return
 }
